BugFix at convertion function"number2hexstring"
Former-commit-id: 68ab1ed11c3765d998f10fa17570d7109c218df8

+++ unisim_lib/unisim/kernel/service/service.hh
@@ -294,8 +294,8 @@ public:
 	virtual double GetHostTime()	{ return (0);	}
 
 	// ********** The following methods are added by Reda and used by PIMServer *****
-	virtual long   GetStructuredAddress(long logicalAddress) { return (logicalAddress); }
-	virtual long   GetPhysicalAddress(long logicalAddress) { return (logicalAddress); }
+	virtual unsigned long long   GetStructuredAddress(unsigned long long logicalAddress) { return (logicalAddress); }
+	virtual unsigned long long   GetPhysicalAddress(unsigned long long logicalAddress) { return (logicalAddress); }
 	// *******************************************************************************
 
 	bool IsWarningEnabled() const;
